More code for planetary interaction monitoring.
Former-commit-id: 75ddd9cb1ad217c8f6e44297ea022ae08eb56068

+++ EVEMon.Common/Serialization/API/SerializablePlanetaryColony.cs
@@ -12,13 +12,13 @@ namespace EVEMon.Common.Serialization.API
         public string PlanetName { get; set; }
 
         [XmlAttribute("planetTypeID")]
-        public long PlanetTypeID { get; set; }
+        public int PlanetTypeID { get; set; }
 
         [XmlAttribute("planetTypeName")]
         public string PlanetTypeName { get; set; }
 
         [XmlAttribute("solarSystemID")]
-        public long SolarSystemID { get; set; }
+        public int SolarSystemID { get; set; }
 
         [XmlAttribute("solarSystemName")]
         public string SolarSystemName { get; set; }
